GIAIC-Learn-Agentic-AI-/Lit-LLM/pyproject.toml
[project]
name = "lit-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
